pyproject.toml
[project]
name = "jam-cot-experiment"
version = "0.1.0"
description = "JAM-CoT prompt tuning experiment on GSM8K"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.0",
    "wandb>=0.16.0",
    "openai>=1.0.0",
    "datasets>=2.14.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"